Adding dockerfiles so far: pybase and arch-pybase.
pybase is based on alpine and builds an image that has pyenv, python 3,
and poetry for doing python dev on top of. arch-pybase was a simple test
to see how much smaller alpine + python tools is compared to archlinux +
python tools. pybase is still much smaller so it will be preferred for
now.